Fixed issue with dragon altar not properly setting altar state
