+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/gui/widgets/buttons/AltarTypeButton.java
@@ -1,17 +1,12 @@
 package by.dragonsurvivalteam.dragonsurvival.client.gui.widgets.buttons;
 
-import by.dragonsurvivalteam.dragonsurvival.DragonSurvivalMod;
-import by.dragonsurvivalteam.dragonsurvival.client.gui.DragonAltarGUI;
+import by.dragonsurvivalteam.dragonsurvival.client.gui.DragonAltarScreen;
 import by.dragonsurvivalteam.dragonsurvival.client.gui.dragon_editor.DragonEditorScreen;
 import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateProvider;
 import by.dragonsurvivalteam.dragonsurvival.common.dragon_types.AbstractDragonType;
 import by.dragonsurvivalteam.dragonsurvival.common.dragon_types.DragonTypes;
 import by.dragonsurvivalteam.dragonsurvival.common.handlers.DragonFoodHandler;
 import by.dragonsurvivalteam.dragonsurvival.config.ServerConfig;
-import by.dragonsurvivalteam.dragonsurvival.network.client.ClientProxy;
-import by.dragonsurvivalteam.dragonsurvival.network.flight.SyncSpinStatus;
-import by.dragonsurvivalteam.dragonsurvival.network.player.SyncDragonHandler;
-import by.dragonsurvivalteam.dragonsurvival.network.status.SyncAltarCooldown;
 import by.dragonsurvivalteam.dragonsurvival.network.syncing.SyncComplete;
 import by.dragonsurvivalteam.dragonsurvival.util.DragonUtils;
 import by.dragonsurvivalteam.dragonsurvival.util.Functions;
@@ -40,10 +35,10 @@ import static by.dragonsurvivalteam.dragonsurvival.DragonSurvivalMod.MODID;
 
 public class AltarTypeButton extends Button {
 	private static final ResourceLocation BACKGROUND_TEXTURE = ResourceLocation.fromNamespaceAndPath(MODID, "textures/gui/dragon_altar_icons.png");
-	private final DragonAltarGUI gui;
+	private final DragonAltarScreen gui;
 	public AbstractDragonType type;
 
-	public AltarTypeButton(DragonAltarGUI gui, AbstractDragonType type, int x, int y){
+	public AltarTypeButton(DragonAltarScreen gui, AbstractDragonType type, int x, int y){
 		super(x, y, 49, 147, Component.empty(), Button::onPress, DEFAULT_NARRATION);
 		this.gui = gui;
 		this.type = type;
