+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/entity/dragon/DragonGlowLayerRenderer.java
@@ -1,7 +1,6 @@
 package by.dragonsurvivalteam.dragonsurvival.client.render.entity.dragon;
 
-import by.dragonsurvivalteam.dragonsurvival.DragonSurvivalMod;
-import by.dragonsurvivalteam.dragonsurvival.client.render.ClientDragonRender;
+import by.dragonsurvivalteam.dragonsurvival.client.render.ClientDragonRenderer;
 import by.dragonsurvivalteam.dragonsurvival.client.skin_editor_system.objects.SkinPreset;
 import by.dragonsurvivalteam.dragonsurvival.client.skin_editor_system.objects.SkinPreset.SkinAgeGroup;
 import by.dragonsurvivalteam.dragonsurvival.client.skins.DragonSkins;
@@ -41,7 +40,7 @@ public class DragonGlowLayerRenderer extends GeoRenderLayer<DragonEntity> {
 			return;
 		}
 
-		if (animatable == ClientDragonRender.dragonArmor) {
+		if (animatable == ClientDragonRenderer.dragonArmor) {
 			return;
 		}
