+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/render/entity/dragon/DragonRenderer.java
@@ -1,6 +1,6 @@
 package by.dragonsurvivalteam.dragonsurvival.client.render.entity.dragon;
 
-import by.dragonsurvivalteam.dragonsurvival.client.render.ClientDragonRender;
+import by.dragonsurvivalteam.dragonsurvival.client.render.ClientDragonRenderer;
 import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateHandler;
 import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateProvider;
 import by.dragonsurvivalteam.dragonsurvival.common.entity.DragonEntity;
@@ -15,14 +15,11 @@ import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemDisplayContext;
-import net.neoforged.fml.ModList;
 import software.bernie.geckolib.cache.object.BakedGeoModel;
 import software.bernie.geckolib.cache.object.GeoBone;
 import software.bernie.geckolib.model.GeoModel;
 import software.bernie.geckolib.renderer.GeoEntityRenderer;
 import software.bernie.geckolib.util.Color;
-import software.bernie.geckolib.util.RenderUtil;
 
 public class DragonRenderer extends GeoEntityRenderer<DragonEntity> {
 	@ConfigOption( side = ConfigSide.CLIENT, key = "renderHeldItem", comment = "Should items be rendered in third person for dragon players?", category = "rendering" )
@@ -33,7 +30,7 @@ public class DragonRenderer extends GeoEntityRenderer<DragonEntity> {
 	public boolean shouldRenderLayers = true;
 	public boolean isRenderLayers = false;
 
-	/** Used when rendering dyeable armor pieces in {@link ClientDragonRender#renderArmorPiece} */
+	/** Used when rendering dyeable armor pieces in {@link ClientDragonRenderer#renderArmorPiece} */
 	public Color renderColor = Color.ofRGB(255, 255, 255);
 
 	public DragonRenderer(final EntityRendererProvider.Context context, final GeoModel<DragonEntity> model) {
@@ -43,9 +40,9 @@ public class DragonRenderer extends GeoEntityRenderer<DragonEntity> {
 		getRenderLayers().add(new ClawsAndTeethRenderLayer(this));
 		getRenderLayers().add(new DragonArmorRenderLayer(this));
 		getRenderLayers().add(new DragonItemRenderLayer(this, (bone, animatable) -> {
-			if(bone.getName().equals(ClientDragonRender.renderItemsInMouth ? "RightItem_jaw" : "RightItem")) {
+			if(bone.getName().equals(ClientDragonRenderer.renderItemsInMouth ? "RightItem_jaw" : "RightItem")) {
 				return animatable.getMainHandItem();
-			} else if(bone.getName().equals(ClientDragonRender.renderItemsInMouth ? "LeftItem_jaw" : "LeftItem")) {
+			} else if(bone.getName().equals(ClientDragonRenderer.renderItemsInMouth ? "LeftItem_jaw" : "LeftItem")) {
 				return animatable.getOffhandItem();
 			}
 			return null;
@@ -82,10 +79,10 @@ public class DragonRenderer extends GeoEntityRenderer<DragonEntity> {
 		if (handler.getBody() != null)
 			hasWings = hasWings || !handler.getBody().canHideWings();
 
-		GeoBone leftWing = ClientDragonRender.dragonModel.getAnimationProcessor().getBone("WingLeft");
-		GeoBone rightWing = ClientDragonRender.dragonModel.getAnimationProcessor().getBone("WingRight");
-		GeoBone smallLeftWing = ClientDragonRender.dragonModel.getAnimationProcessor().getBone("SmallWingLeft");
-		GeoBone smallRightWing = ClientDragonRender.dragonModel.getAnimationProcessor().getBone("SmallWingRight");
+		GeoBone leftWing = ClientDragonRenderer.dragonModel.getAnimationProcessor().getBone("WingLeft");
+		GeoBone rightWing = ClientDragonRenderer.dragonModel.getAnimationProcessor().getBone("WingRight");
+		GeoBone smallLeftWing = ClientDragonRenderer.dragonModel.getAnimationProcessor().getBone("SmallWingLeft");
+		GeoBone smallRightWing = ClientDragonRenderer.dragonModel.getAnimationProcessor().getBone("SmallWingRight");
 
 		if (leftWing != null)
 			leftWing.setHidden(!hasWings);
